Refactor structure validation logic to utilize a dedicated subsection validator; improve error handling and reduce code duplication
